Carry `k` as a specta-exportable i32 and regenerate the bindings
F4 widened `k` from `Option<u8>` to `Option<i64>` to stop `-1`/`256` dying in
transport decoding — the right instinct, but `i64` is a type specta REFUSES to
export (the `BigInt`-style integers: `i64`/`u64`/`usize`/`isize`/`i128`/`u128`
risk precision loss crossing as a TS `number`). So the failure moved rather than
went away: `cargo run --example export-bindings` failed outright on
`WalkForwardRunRequest.k`, which made `scripts/check-bindings.sh` fail at
"Gate -- bindings freshness" on every head from `5fda1a4` onward — a required
check, red on a surface no Rust test touches.

`k` is now `Option<i32>` on all three carriers — the Tauri
`WalkForwardRunRequest`, the MCP `RunWalkForwardArgs` (parity: both transports
carry one wire type for `k`) and the application's `WalkForwardRequest` — and the
application widens once to `i64` for `FoldScheme::rolling_oos`, so the domain's
checked narrowing to `u8` and its typed `KOutOfRange` refusal are unchanged.
`i32` is the widest signed type specta exports, and it is still wide enough to
carry `-1` and `256` to that refusal instead of failing in decoding — which is
the whole point of F4, now preserved on both surfaces rather than one.

`ui/src/bindings.ts` is regenerated with the repo's own mechanism
(`cargo run --quiet --example export-bindings -- ui/src/bindings.ts`) and
committed: the gate's contract is that the committed artifact matches a fresh
generation, so a wire change without it IS this failure. The regeneration also
picks up the N1 doc text on `LibraryVersion.stats` that the failing exporter had
been masking — a second staleness the same gate would have caught once it could
run at all.

The two transports' existing F4 pins still hold and were re-run by name: the
`-1`/`256` cases refuse as a `BusError` with `code = validation` whose message
names `` `k` `` on Tauri, and as a `field_error` on `field = "k"` on MCP — the
typed refusals, not decode failures. The stale "the wire value is i64" comment in
the Tauri pin is corrected to `i32`.

Verified: `bash scripts/check-bindings.sh` OK (1284 lines); all eleven
`scripts/check-*.sh` gate scripts OK; `cargo fmt --check` clean;
`cargo clippy --all-targets -- -D warnings` clean; the affected binaries
`tauri_walk_forward` / `mcp_walk_forward` / `walk_forward` 27/27 and the three F4
pins 3/3; `cargo nextest run --no-fail-fast` 1182 passed / 2 skipped / 1 failed
where the failure is issue #177, the known `coach_turn_boundary` flake (same
panic), green on isolated retry. The UI half of the same check also passes:
`npm run typecheck` clean and `npm run test` 154 passed / 10 files.

+++ src/mcp/tools.rs
@@ -194,8 +194,13 @@ pub(crate) struct RunWalkForwardArgs {
     /// The fold count — `2..=12`; omitted defaults to 6. Wider than the legal
     /// `u8` on purpose: an out-of-range value must reach the domain's
     /// `KOutOfRange` refusal (a `field_error` naming `k`), not die in decoding.
+    ///
+    /// `i32`, matching the Tauri transport: the two surfaces carry one wire type
+    /// for `k`, and it is the widest signed one specta exports (the `BigInt`-style
+    /// integers are refused, so `i64` would break the bindings export rather than
+    /// this decode).
     #[serde(default)]
-    k: Option<i64>,
+    k: Option<i32>,
 }
 
 /// `get_walk_forward_run` args (r2.s3.w5).
